[WFLY-11813] Added test for empty username for Elytron's FormAuthenticationMechanism + additional minor refactoring
